pipelines/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
apache-beam==2.63.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "linux" or platform_python_implementation == "PyPy")
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
cloudpickle==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
crcmod==1.7 ; python_version >= "3.12" and python_version < "4.0"
cryptography==44.0.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "linux"
deprecated==1.2.18 ; python_version >= "3.12" and python_version < "4.0"
dill==0.3.1.1 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
docopt==0.6.2 ; python_version >= "3.12" and python_version < "4.0"
docstring-parser==0.16 ; python_version >= "3.12" and python_version < "4.0"
fastavro==1.10.0 ; python_version >= "3.12" and python_version < "4.0"
fasteners==0.19 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-apitools==0.5.31 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.38.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-aiplatform==1.84.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery-storage==2.29.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery==3.30.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigtable==2.29.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-datastore==2.20.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-dlp==3.28.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-language==2.17.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-pubsub==2.28.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-pubsublite==1.12.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-recommendations-ai==0.10.16 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-resource-manager==1.14.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-spanner==3.53.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==2.19.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-videointelligence==2.16.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-vision==3.10.0 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.69.1 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.14.1 ; python_version >= "3.12" and python_version < "4.0"
grpc-interceptor==0.15.4 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.65.5 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.65.5 ; python_version >= "3.12" and python_version < "4.0"
hdfs==2.7.3 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.12" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
jaraco-context==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
jaraco-functools==4.1.0 ; python_version >= "3.12" and python_version < "4.0"
jeepney==0.9.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "linux"
jsonpickle==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.12" and python_version < "4.0"
keyring==25.6.0 ; python_version >= "3.12" and python_version < "4.0"
keyrings-google-artifactregistry-auth==1.1.2 ; python_version >= "3.12" and python_version < "4.0"
more-itertools==10.6.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
oauth2client==4.1.3 ; python_version >= "3.12" and python_version < "4.0"
objsize==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.31.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-sdk==1.31.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-semantic-conventions==0.52b0 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.15 ; python_version >= "3.12" and python_version < "4.0"
overrides==7.7.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.3 ; python_version >= "3.12" and python_version < "4.0"
pyarrow-hotfix==0.6 ; python_version >= "3.12" and python_version < "4.0"
pyarrow==16.1.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "linux" or platform_python_implementation == "PyPy")
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
pydot==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
pymongo==4.11.2 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.12" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.12" and python_version < "4.0"
referencing==0.36.2 ; python_version >= "3.12" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rpds-py==0.23.1 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "linux"
shapely==2.0.7 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.17.2 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
zstandard==0.23.0 ; python_version >= "3.12" and python_version < "4.0"
